amend! refactor: reduce time password stays in-memory after onboarding
refactor: reduce time password stays in-memory after onboarding

Add a new useSecrets hook and use it in useWalletManager instead of
passing password around as prop.

Explicitly set html input[type=password].value to '' after onboarding

Hoist Password and PasswordVerification components from 'common' to
'core' package and update them to use the new UncontrolledPasswordBox
component from the UI Toolkit.
