+++ apps/browser-extension-wallet/src/features/dapp/components/SignTransaction.tsx
@@ -2,7 +2,8 @@ import React, { useCallback, useState, useMemo } from 'react';
 import { Spin } from 'antd';
 import { Wallet } from '@lace/cardano';
 import { useTranslation } from 'react-i18next';
-import { Button, OnPasswordChange, Password, PostHogAction } from '@lace/common';
+import { Button, PostHogAction } from '@lace/common';
+import { OnPasswordChange, Password } from '@lace/core';
 import { useRedirection } from '@hooks';
 import { dAppRoutePaths } from '@routes';
 import { Layout } from './Layout';
