+++ apps/browser-extension-wallet/src/features/unlock-wallet/components/UnlockWallet.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { ContentLayout } from '@components/Layout';
 import laceLogoMark from '@src/assets/branding/lace-logo-mark.svg';
-import { Button, Password, PasswordProps } from '@lace/common';
+import { Button } from '@lace/common';
+import { Password, PasswordProps } from '@lace/core';
 import styles from './UnlockWallet.module.scss';
 import { useTranslation } from 'react-i18next';
 import { ForgotPassword } from './ForgotPassword';
