+++ apps/browser-extension-wallet/src/features/unlock-wallet/components/UnlockWalletContainer.tsx
@@ -4,7 +4,8 @@ import { useWalletManager } from '@src/hooks/useWalletManager';
 import { useWalletStore } from '@src/stores';
 import { useBackgroundServiceAPIContext } from '@providers/BackgroundServiceAPI';
 import { saveValueInLocalStorage } from '@src/utils/local-storage';
-import { OnPasswordChange, useKeyboardShortcut } from '@lace/common';
+import { useKeyboardShortcut } from '@lace/common';
+import { OnPasswordChange } from '@lace/core';
 import { BrowserViewSections } from '@lib/scripts/types';
 import { useAnalyticsContext } from '@providers';
 import { PostHogAction } from '@providers/AnalyticsProvider/analyticsTracker';
