--- packages/common/src/ui/components/Form/Password/index.ts
@@ -1,1 +1,0 @@
-export * from './Password';
