+++ packages/common/src/ui/components/Form/index.tsx
@@ -1,4 +1,2 @@
 export * from './Input';
 export * from './TextArea';
-export * from './Password';
-export * from './PasswordVerification';
