+++ packages/core/src/ui/components/Password/Password.module.scss
