+++ packages/core/src/ui/components/Password/index.ts
@@ -1,0 +1,2 @@
+export * from './Password';
+export type { Password as PasswordObj, OnPasswordChange } from '@input-output-hk/lace-ui-toolkit';
