+++ packages/core/src/ui/components/PasswordVerification/PasswordVerification.module.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/theme.scss';
-@import '../../../styles/abstracts/typography';
+@import '../../../../../common/src/ui/styles/theme.scss';
+@import '../../../../../common/src/ui/styles/abstracts/typography';
 
 .container {
   display: flex;
