+++ packages/core/src/ui/components/WalletSetup/WalletSetupNamePasswordStep/types.ts
@@ -1,4 +1,4 @@
-import { PasswordVerificationProps } from '@lace/common';
+import { PasswordVerificationProps } from '@ui/components/PasswordVerification';
 
 export type ValidationErrorKeys = 'nameMaxLength';
 export type BarStates = PasswordVerificationProps['complexityBarList'];
